+++ src/main.rs
@@ -1,11 +1,11 @@
 extern crate clap;
 
 use clap::{arg, value_parser, Arg, ArgAction, Command};
-use coarsetime::Duration;
 use image::Luma;
 use qrcode::QrCode;
 use std::num::ParseIntError;
 use std::path::PathBuf;
+use std::time::Duration;
 
 fn cli() -> Command {
     Command::new("xqr")
@@ -21,7 +21,7 @@ fn cli() -> Command {
                         .value_parser(value_parser!(PathBuf))
                         .required(true),
                 )
-                .arg(arg!(kid: --kid <KID> "The key ID to use").required(true))
+                .arg(arg!(iss: --iss <ISS> "The issuer to use").required(true))
                 .arg(
                     Arg::new("display")
                         .long("display")
@@ -72,10 +72,10 @@ fn main() {
                 }
             };
             let value = sub_matches.get_one::<String>("value").unwrap();
-            let kid = sub_matches.get_one::<String>("kid").unwrap();
+            let iss = sub_matches.get_one::<String>("iss").unwrap();
             let valid_for = sub_matches.get_one::<Duration>("valid_for");
 
-            match xqr::encode(&priv_key, value, kid, valid_for.copied()) {
+            match xqr::encode(&priv_key, value, iss, valid_for.copied()) {
                 Ok(encoded_xqr) => {
                     if sub_matches.get_flag("display") {
                         match qr2term::print_qr(&encoded_xqr.token) {
@@ -127,14 +127,21 @@ fn main() {
                     std::process::exit(1);
                 }
             };
-            let pub_key = match xqr::fetch_public_key(&kid) {
-                Ok(pub_key) => pub_key,
-                Err(e) => {
-                    eprintln!("Error fetching public key for key id ({}): {}", kid, e);
+            let iss = match encoded_xqr.get_iss() {
+                Some(iss) => iss,
+                None => {
+                    eprintln!("Error decoding: {}", "Issuer not found in token");
                     std::process::exit(1);
                 }
             };
-            match xqr::decode(&pub_key, encoded_xqr) {
+            let pub_key = match xqr::fetch_public_key(&iss, &kid) {
+                Ok(pub_key) => pub_key,
+                Err(e) => {
+                    eprintln!("Error fetching public key for key id ({}): {}", iss, e);
+                    std::process::exit(1);
+                }
+            };
+            match xqr::decode(&pub_key, &encoded_xqr) {
                 Ok(decoded_value) => println!("{}", decoded_value),
                 Err(e) => {
                     eprintln!("Error decoding: {}", e);
@@ -145,15 +152,21 @@ fn main() {
         Some(("generate-key-pair", sub_matches)) => {
             let save_path = sub_matches.get_one::<PathBuf>("save").unwrap();
             let pub_key_path = save_path.with_extension("pub");
-            let key_pair = xqr::generate_key_pair();
-            let private_key_pem = match key_pair.to_pem() {
+            let priv_key = match xqr::generate_key() {
+                Ok(k) => k,
+                Err(e) => {
+                    eprintln!("Error generating key pair: {}", e);
+                    std::process::exit(1);
+                }
+            };
+            let private_key_pem = match priv_key.private_key_to_pem_pkcs8() {
                 Ok(private_key_pem) => private_key_pem,
                 Err(e) => {
                     eprintln!("Error generating private key: {}", e);
                     std::process::exit(1);
                 }
             };
-            let pub_key_pem = match key_pair.public_key().to_pem() {
+            let pub_key_pem = match priv_key.public_key_to_pem() {
                 Ok(pub_key_pem) => pub_key_pem,
                 Err(e) => {
                     eprintln!("Error generating public key: {}", e);